agent/requirements.txt
# FastAPI dan dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
mysql-connector-python==8.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Computer Vision
ultralytics==8.0.206
opencv-python==4.8.1.78
Pillow==10.1.0

# NLP & API calls
groq==0.4.1
requests==2.31.0

# Data processing
numpy==1.25.2
pandas==2.1.3

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
